requirements.txt
alembic==1.4.2
Babel==2.8.0
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
flask-blacklist==0.0.1
flask-praetorian==1.0.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
Flask-Cors
passlib==1.7.2
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pytzdata==2019.3
SQLAlchemy==1.3.16
Werkzeug==1.0.1
wrapt==1.12.1